Fix disconnect handling and advancing criteria
The case where a player disconnects wasn't being handled properly for
lockstep mode: the sync layer (sagely) doesn't allow 'rolling back' to
the current frame, but in the case of a player disconnecting then it is
possible to determine an 'incorrect' frame of the current frame. We
handle that by simply disregarding "incorrect frames" that are equal to
the current frame.

Additionally, the logic for whether to advance frames was quite
convoluted and probably incorrectly handling the case where no frames
were yet confirmed, so this also clarifies the logic there.