Fix desync detection false positive caused by a rollback
This fixes the case where the following happens:

1. Sync layer requests roll back to frame N
2. Sync layer requests advance to frame N+1
3. Sync layer marks frame N as confirmed
4. P2pSession's desync detection decides to send a checksum for frame N
   to peers, even though the checksum for frame N is now known to be incorrect,
   (because ggrs-consuming-code has not had a chance to act on the
   advance frame request, so the checksum stored in frame N is from
   before the rollback.)
5. Peers receive the incorrect checksum and flip out thinking there is a
   desync.

Actually causing this bug to reliably manifest in practice is somewhat
tricky; I was able to reliably do it in my game via the following:

* high input latencies (150ms on each client network interface, so 300ms
  round trip)
* input delay == 0
* desync detection interval set to 1
* start all game clients at once
* in the first advance, each client initially sends an input which is
  mispredicted by all other clients (so not the zeroed input).
* and lastly that first frame from each client must modify the resulting
  checksum in a way that is unique from each other client.

The concrete input I am sending in my game is approximately equivalent
to "store the client handle in the game state", but but when I tried
modifying the example game to do similar then it didn't manifest this
bug - no idea why. In any case, this patch fixes the issue so I'll leave
it at that.

It also changes a < to <= in a seemingly-isolated way; this change is
based on the rationale given at
https://github.com/gschup/ggrs/pull/64/files#r1368265924 no longer
applying.