WIP: Allow player handles to have gaps in them (spectating broken)
* Player handles are now newtyped
* Maintaining explicit mappings between player handles and their state
  (rather than using the player handle as an index into a Vec)
  throughout the code makes it easier to understand the dependencies on
  players.
* Player handles can now be discontiguous, which enables a session to be
  recreated from a past session's state while maintaining consistent
  PlayerHandles for players that are still around. This is also a key
  step towards implementing hot join.

However, note that the logic for sending and receiving Input messages
still expects both sender and receiver to have identical sets of
PlayerHandles in their sessions; fixing this would require including the
PlayerHandles in the Input message so that they can be mapped to the
input.

Lastly, in this changeset spectating is currently broken.